requirements.txt
protobuf==5.26.1
streamlit==1.34.0
streamlit_authenticator==0.1.5
streamlit_option_menu==0.3.12
google-generativeai==0.5.2